Fix sync issues with best block
