+++ src/plugins/explorer/events.js
@@ -1,0 +1,12 @@
+'use strict'
+
+function createEventsRegistry () {
+  const registeredEvents = []
+
+  return {
+    getAll: () => registeredEvents,
+    register: registration => registeredEvents.push(registration)
+  }
+}
+
+module.exports = createEventsRegistry
